Move custom exception definitions into a submodule
